+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Make paint dependency optional
 - Remove pause feature
 - Separate configuring into its own method, remember whirly's configuration, can be cleared with the new .reset method
+- Introduce "stop" frames to display when spinner is over
 
 ### 0.1.1